Updated Base Models
I've updated the base models to follow the comments.

Remarks:
- I placed the Difficulty enum in its own folder called "Enums" which may be weird, I'm not sure

- Concretions of ToolBase will now also have to implement either IXPCost or ICoinCost, as ToolBase itself does not implement either.

- Looking at IXPCost and ICoinCost, we can probably do the same for Income. Projects and Jobs have income in coins, and this property is in both ProjectBase and JobBase, we could also say that Languages/Tools/Frameworks have income in XP (although not player XP).
So when you complete a learning session -> XP in that language/framework/tool.
When you complete a job/project session --> coins.
We could have this as IXPIncome and ICoinIncome, similar to IXPCost and ICoinCost.

- Grouping the interfaces? The project may get cluttered as we add more interfaces, do you think we should group interfaces in folders?
Eg: The cost and income interfaces, or just all interfaces in one folder.